requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

flask
connexion[swagger-ui]
connexion[flask]
connexion[uvicorn]
numpy
opencv-python==4.5.5.62
pillow
PyYAML
timm==0.4.12
torchvision==0.11.3
torch==1.10.2
transformers==4.8.1
